@@ -58,10 +58,11 @@
     "leaflet-rotatedmarker": "^0.2.0",
     "lodash": "^4.16.4",
     "moment": "^2.15.1",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "prop-types": "^15.7.2",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
     "react-redux": "^5.0.6",
-    "react-router-dom": "^4.2.2",
+    "react-router-dom": "^4.3.1",
     "redux": "^3.7.2"
   },
   "scripts": {
